+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 Only listing significant user-visible, not internal code cleanups and minor bug fixes.
 
+## 0.7.0 (upcoming)
+
+* Pending changelog
+
 ## 0.6.0 (upcoming)
 
 * [QATM-236] Quit after tagged scenario fails